third_party/blink/web_tests/paint/invalidation/svg/viewport-mask-update-expected.txt

CONSOLE ERROR: line 8: Uncaught SyntaxError: Unexpected token ','
CONSOLE ERROR: line 11: Uncaught ReferenceError: repaintTest is not defined
